+++ packages/terra-framework-docs/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Added
+  * Added example to `terra-table` to show column states and sorting;
+  
 * Updated
   * Updated About page of `terra-table` to provide accessibility documentation for Home and End keys.